+++ src/hooks/useUrlParams.ts
@@ -1,4 +1,4 @@
 import { useParams } from 'react-router-dom';
-import UrlParams from '../routing/ulr-params';
+import UrlParams from '../routing/url-params';
 
 export const useUrlParams = (): UrlParams => useParams<UrlParams>();
