+++ src/routing/user/user.tsx
@@ -3,7 +3,7 @@ import { Route, Switch, useRouteMatch } from 'react-router-dom';
 import { FourOuFour } from '../../pages';
 import EditUserProfilePage from '../../pages/User/EditUserProfilePage';
 import { UserProfilePage } from '../../pages/User/UserProfilePage';
-import { nameOfUrl } from '../ulr-params';
+import { nameOfUrl } from '../url-params';
 
 export const UserRoute: FC = () => {
   const { path } = useRouteMatch();
